+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unversioned
 
 - Minor: Added an option to allow mass pings while the streamer is on/offline. (#1129)
+- Minor: Added an option to allow rouletting while the streamer is on/offline. (#1131)
 - Minor: Added `$(date:<timezone>)` variable. (#1125)
 
 ## v1.48